requirements.txt
streamlit>=1.30

# RAG stack
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.20
langchain-text-splitters>=0.0.1

# Vector DB
chromadb>=0.4.22

# OpenAI-compatible client used by LangChain ChatOpenAI/OpenAIEmbeddings
openai>=1.0.0

# Optional (used by your scripts; safe to include)
python-dotenv>=1.0.0
pymupdf>=1.23.0

